OXDEV-2641 - cleanup acceptance testing
